+++ c1/Libft/ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: sersanch <sersanch@student.42barcelona.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/15 12:00:51 by sersanch          #+#    #+#             */
-/*   Updated: 2022/09/24 11:46:19 by sersanch         ###   ########.fr       */
+/*   Updated: 2022/09/27 15:16:17 by sersanch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,16 +14,16 @@
 
 void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
-	int		i;
-	char	*dstc;
-	char	*srcc;
+	unsigned int	i;
+	char			*dstc;
+	char			*srcc;
 
 	if (!dst && !src && n > 0)
 		return (0);
 	dstc = (char *)dst;
 	srcc = (char *)src;
 	i = 0;
-	while (i < n)
+	while (i < (unsigned int)n)
 	{
 		dstc[i] = srcc[i];
 		i++;
