+++ c1/Libft/ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: sersanch <sersanch@student.42barcelona.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/19 10:33:07 by sersanch          #+#    #+#             */
-/*   Updated: 2022/09/19 14:27:35 by sersanch         ###   ########.fr       */
+/*   Updated: 2022/09/27 15:17:17 by sersanch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 	dst_count = ft_strlen(dst);
 	if (dstsize == 0)
 		return (src_count);
-	if (dstsize < dst_count)
+	if (dstsize < (size_t)dst_count)
 		src_count += dstsize;
 	else
 		src_count += dst_count;
